requirements.txt
huggingface_hub>=0.34.0
python-dotenv>=1.0.1
pypdf>=4.0.0
fastembed>=0.5.0
